+++ src/main/index.ts
@@ -154,9 +154,9 @@ export function thriftProjectFromSourceFiles(
         {},
     )
 
-    const resolvedInvalidFiles: Array<INamespace> = Utils.collectInvalidFiles(
-        Utils.valuesForObject(resolvedNamespaces),
-    )
+    const resolvedInvalidFiles: Array<
+        INamespace
+    > = Debugger.collectInvalidFiles(Utils.valuesForObject(resolvedNamespaces))
 
     if (resolvedInvalidFiles.length > 0) {
         Debugger.printErrors(resolvedInvalidFiles)
@@ -172,7 +172,7 @@ export function thriftProjectFromSourceFiles(
 
         const validatedInvalidFiles: Array<
             INamespace
-        > = Utils.collectInvalidFiles(
+        > = Debugger.collectInvalidFiles(
             Utils.valuesForObject(validatedNamespaces),
         )
