+++ src/main/render/thrift-server/struct/encode.ts
@@ -15,7 +15,7 @@ import { COMMON_IDENTIFIERS, THRIFT_IDENTIFIERS } from '../identifiers'
 
 import { WRITE_METHODS, WriteMethodName } from './methods'
 
-import { resolveIdentifierDefinition } from '../../../resolver'
+import { Resolver } from '../../../resolver'
 
 import {
     coerceType,
@@ -241,7 +241,7 @@ export function writeValueForType(
 ): Array<ts.Expression> {
     switch (fieldType.type) {
         case SyntaxType.Identifier:
-            const definition: DefinitionType = resolveIdentifierDefinition(
+            const definition: DefinitionType = Resolver.resolveIdentifierDefinition(
                 fieldType,
                 {
                     currentNamespace: state.currentNamespace,
