+++ src/main/render/thrift-server/struct/reader.ts
@@ -16,7 +16,7 @@ import {
 
 import { className, toolkitName } from './utils'
 
-import { resolveIdentifierDefinition } from '../../../resolver'
+import { Resolver } from '../../../resolver'
 import { DefinitionType, IRenderState } from '../../../types'
 import { createMethodCall } from '../../shared/utils'
 import { COMMON_IDENTIFIERS } from '../identifiers'
@@ -150,7 +150,7 @@ export function assignmentForFieldType(
 ): Array<ts.Statement> {
     switch (fieldType.type) {
         case SyntaxType.Identifier:
-            const definition = resolveIdentifierDefinition(fieldType, {
+            const definition = Resolver.resolveIdentifierDefinition(fieldType, {
                 currentNamespace: state.currentNamespace,
                 namespaceMap: state.project.namespaces,
             })
