+++ tsconfig.test.json
@@ -9,7 +9,7 @@
         "outDir": "./dist",
         "noEmitOnError": true,
         "strict": true,
-        "noUnusedLocals": false,
+        "noUnusedLocals": true,
         "pretty": true,
         "removeComments": true
     },
